Initial commit for cascade-select custom element
This initial commit includes the implementation of the cascade-select custom element.
The index.html serves as an example on how to use CascadeSelect. The index.js file defines the CascadeSelect class which extends the HTMLElement class. It implements the connectedCallback, disconnectedCallback, collectOptions, createSelects, appendToShadowRoot, createSelect, updateOptions, and createSelectOptions methods. The updateOptions method updates the options in the second select based on the selected value in the first select element. The customElements.define is used to register a custom element.